10/10 pylint, 90%+ coverage, added program and test to create csv files
