+++ students/dfspray/Lesson05/assignment/src/csvs/customer_data.csv
@@ -1,0 +1,4 @@
+id,name,address,phone_number
+cst001,Charlie,123 Fleet Street,1231234123
+cst002,Andrey,123 Jordan Street,4564564567
+cst003,Vijay,123 Lake Street,7987897891
