+++ students/dfspray/Lesson05/assignment/src/csvs/product_data.csv
@@ -1,0 +1,4 @@
+id,description,product_type,quantity_available
+prd001,TV,livingroom,3
+prd002,Couch,livingroom,1
+prd003,Chair,livingroom,0
