+++ students/dfspray/Lesson05/assignment/src/csvs/rentals_data.csv
@@ -1,0 +1,4 @@
+id,name,rentals
+cst001,Charlie,prd001 prd002
+cst002,Andrey,prd002 prd003
+cst003,Vijay,prd001 prd003
